代码构造优化 pixiv数据不再依赖第三方接口，改为爬虫获取

+++ src/main/java/gugugu/apirequest/imgsearch/PixivImjadIllustGet.java
@@ -29,7 +29,7 @@ public class PixivImjadIllustGet extends BaseRequest {
     /**
      * p站图片id
      */
-    private int pixivId;
+    private Long pixivId;
 
     //执行接口请求
     public void doRequest() throws IOException {
